requirements.txt
streamlit
python-dotenv
sqlite-vec
google-genai
passlib
argon2
sentence_transformers
python-dotenv